Support non-uniform categoricals (#1877)
Add support for non-uniform categoricals.

This adds support for different partitions having different categorical
mappings. This is accomplished through `union_categoricals`, a method added in
pandas 0.19.0. As such, support for pandas < 0.19.0 has been dropped.

### Changes to metadata:

Categorical columns/indices now have two states:

- known categoricals have the `categories` known statically (on the `_meta`
  attribute)
- unknown categoricals don't know the categories statically, and are indicated
  by the presence of `dd.utils.UNKNOWN_CATEGORIES` in the categories on the
  meta attribute. This is nice, because operations on the categorical in most
  cases propagate the presence of this value, so the known/unknown state
  carries through inference.

### Changes to the categorical accessor

To query and convert between these states, following has been added to the
categorical accessor:

- `.cat.known` is a property that indicates if the categories are known
- `.cat.as_known()` converts unknown to known (and is a no-op if already
  known). Note that this requires computation.
- `.cat.as_unknown()` converts known to unknown. This requires no computation -
  it's just a change in metadata.

If the categories are known, then `.cat.categories` returns statically (no
computation needed). If they are unknown, then an error is raised (rather than
computing them)

### Changes to concat

Concatenation of *dask* dataframes with different categoricals is now
supported. For each column/index in a dataframe, the output *dask* column/index
will have known categories only if all categories are known. Otherwise the
categories will be unknown.

After a call to concat, each partition is guaranteed to match the metadata on
the containing dask dataframe. This fixes a bug from before where in certain
cases partitions could have different columns/datatypes.

### Changes to `pivot_table` and `get_dummies`

These functions only work if the categories are known. If unknown, an error is
raised

### Changes to `read_csv`

For pandas > 0.19.2, dask now supports reading from csvs directly into
categorical columns. These will be unknown categories

### Changes to `merge`

No changes are made to merge. Pandas merge doesn't support merging
categoricals, *even if the categoricals are identical*. As such, dask doesn't
error, but the categorical datatype is dropped. When this is fixed in pandas,
support for merge on differing categoricals can be added.

### Changes to `hash_pandas_object`

Previously, this would hash categories based on codes only. We now hash based
on the values the categoricals represent. This was patched in pandas, and will
be in the next release. The fix is backported to dask for versions 0.19.0 to
0.19.2. The faster (and more stable) object hash will be used for pandas
0.19.2.

### Changes to `categorize`

Categorize now takes an optional ``index`` parameter, indicating whether to
categorize the index or not. Additionaly, by default it will convert all
unknown categoricals to known categoricals.