Fix when performing multiple reductions on a sparse bag.
When doing `db.compute(bag.min(), bag.max())`, the iterable on each partition
is converted to a list so it can be passed to both reducers.  Previously,
if `empty_safe_apply` was given a list, it wouldn't return `no_result`.  We
are now explicit about which tasks are the final aggregations.