dask.delayed functions are serializable
Functions are pickled by their global name (combination of ``__module__``
and ``__qualname__``). Since decorating a function in python replaces the
name with the result from the decorator, this can cause problems if you
still want to be able to pickle the underlying function (as we do in
``dask.delayed``).

To deal with this, we wrap all functions in ``DelayedLeaf`` in a proxy
object that supports ``__call__`` and ``__reduce__``. These objects then
serve as callables in the resulting dask graph.

On a single machine, there is negligible overhead (ns increase
per-call). When distributed, the ``__reduce__`` method determines if the
global name of the underlying callable points to a ``DelayedLeaf``,
indicating that it was created by a decorator.  If so, the callable is
serialized to get the global ``DelayedLeaf``, and then extract the
function from the ``_obj`` attribute.  Otherwise, the function is
serialized as normal. This computation is cached, and takes microseconds
on my machine. Note that in the case where the function is serialized
normally, the deserialized result will be just the function, instead of
an instance of ``MaybeDecorated``. This is because we know that the
proxy object isn't needed in these cases, so it's not necessary to
recreate it.