`dask.bag.from_delayed` may create iterators.  Reify to lists if necessary (#1966)
* `dask.bag.from_delayed` may create iterators.  Reify to lists if necessary.

If the `delayed` functions used to create bags return iterators, then we
should convert them to lists iff they are directly used by multiple tasks.
If we don't do this, then the first task will exhaust the iterator, thereby
leaving nothing for the second task to operate on.

`reify` and `list` will be optimized out by the bag optimizations as
appropriate.

* Oups.  That test wasn't needed anyway.

Still, it's interesting that `pluck` in `toolz` and `cytoolz` behave
differently in this exceptional case.  I should fix that.