Keep first finished task in double stealing event (#875)
* Keep first finished task in double stealing event

When a task is stolen during execution we get two copies throughout the
cluster.  Previously if the task had finished by the time the "this task has
been duplicated" message arrived it would delete itself, even if it was the
first to finish (and thus the only to be saved).  This resulted in neither of
the tasks being available.

Now we keep the original task and tell all future tasks to remove themselves.

* release stolen keys if not executing or in memory

* remove workers if we can't gather from them

* gather_from_workers returns more information about how things failed

* Don't release repeated key if this worker holds the key

* cleanup

* cleanup