Experiment with a cheaper task renamer:
a sorted set of keys returned by key_split except for the root node, which is
the last item in the joined keys.  This should be cheap enough to compute,
nice enough to look at, and well-behaved enough to test.  Of course, this
also breaks existing tests that check renamed keys, which will need fixed
if this scheme is approved.