#1964. Fuse tasks that form reductions (aka diamond fuse) (#1979)
* WIP #1964.  Probably broken.  Fuse tasks that form reductions.

This trades parallelism opportunities for faster scheduling by making
tasks less granular.

There are many options (and opinions) for what the parameterization
of this function should be.  For the sake of experimentation, I have
included many parameters as described below.  My gut feeling right
now is that we only need one parameter, `ave_width`, and to choose
a reasonable value of `max_depth_new_edges` based on `ave_width`
(this controls against pathologies while allowing desirable behavior).

This optimization operation is more general than "single input, single
output".  It applies to all reductions, so it may be "multiple input,
single output".  I have attempted to make this well-behaved and
robust against pathologies.  Notably, by allowing multiple inputs
(what I refer to as `edges` in the code), there may be parallelism
opportunities w.r.t. the edges that are otherwise not captured by
analyzing the fusible reduction tasks alone.  I added a cheap-to-
compute heuristic that is conservative; i.e., it will never
underestimate the degree of edge parallelism.  This heuristic
increases the value compared against `min_width`--a measure of
parallizability--which is one of the reasons I prefer `min_width`.

I think it will be easy for this operation to supercede `fuse`.
I'm ignoring task renaming for now.

Parameters
----------
dsk: dict
    dask graph
keys: list or set
    Keys that must remain in the dask graph
ave_width: float
    Limit for `width = num_nodes / height`, a good measure of
    parallelizability.
max_depth_new_edges: int
    Don't fuse if new dependencies are added after this many levels
max_height: int
    Don't fuse more than this many levels
max_width: int
    Don't fuse if total width is greater than this

* Begin adding tests (and a fix).

* Another test, another fix.

I'm beginning to get confused.  I think I should make the traversal strategy clearer.

* Another test, another fix.

I'm still a little confused and should rework things to be more explicit.
Sorry for the tangle of logic with implicit behaviors.

* A little more clear.

* Add a few more tests.  Still not ready for prime time.

* Small tweaks.

* Remove the use of `irreducible` and change `deps` to sets.

We now allow for fusing reducible tasks below tasks that weren't fused, but
only in strict circumstances (such as allowing for linear fusing).

* Remove unnecessary indent, as `num_processing` may only contain positive numbers

* For clarity.  I think.  Goodnight!

* Remove `num_processing` stack, as it no longer provides a benefit.

Time for more tests!

* More reorganization and cleanup.

* Cleaner control flow.  Things are looking better.

* Handle linear fusing so this can replace `fuse` optimization.

This actually performs a superset of linear fusing that `fuse` performs.
Specifically, `fuse_reductions(dsk, ave_width=1)` will fuse linear tasks
that may refer to the inputs used in the top-level task.  In `fuse`, we
only fuse chains with single-input, single-output.

* No more `continue`s or implied fall-through behavior.  More explicit and clear?

I think the clarity and cleanliness is worth the cost of sometimes performing
`item = list.pop()` followed by `list.append(item)`, which is what we avoided
doing previously.

* Accept and return `dependencies` like `fuse` does.  Barely tested.

* Oups

* Optionally get parameters from global context.

* Fix updating `dependencies`, and make it easy to run `fuse` tests.

* Test heuristic, and include dependencies in tests.

Also, test `fuse_reductions` in `fuse` tests.

* Improve the heuristic and track the fudge factor separately.

This is still cheap.  Notably, we don't need to look at and track the
dependencies or dependents of the non-reducible nodes.

* Update docstring and clean up tests.

* Silly micro-optimizations (but, hey, these add up!).  I couldn't help myself.

* Create fast branches when fusing a single child (i.e., linear chains). This is worth it.

* More optimizations.  More aggressive fast path for linear chains.

This did become a little uglier and has a little bit of duplicated code, but
the performance is pretty good!

* The solution to task renaming was to think.  Go brain!

I changed the parameter name from `rename_fused_keys` to `rename_keys`, and
for consistency the global parameter is now `fuse_rename_keys`.  Should we
also change the parameter name of `rename_fused_keys` to `fuse_rename_keys`
in the `optimize` function for the dask collections?

* Experiment with a cheaper task renamer:

a sorted set of keys returned by key_split except for the root node, which is
the last item in the joined keys.  This should be cheap enough to compute,
nice enough to look at, and well-behaved enough to test.  Of course, this
also breaks existing tests that check renamed keys, which will need fixed
if this scheme is approved.

* Experiment with fancy fusing!

* A little better, and basic docstrings.

* Update tests to pass with new renaming method.

* Rename `fuse` to `fuse_linear`, and `fuse_reductions` to `fuse`

* Cleanup.  Remove fancy fuse.