Check and compare shapes in assert_eq
Seems that shape comparison only occurred on the result of computing the
Dask array. In our case, we found the Dask array itself had the wrong
shape, but the computed result had the right shape. To fix this we add a
comparison of the shapes before computation occurs. This happens in
addition to the existing test comparing the shapes afterwards.