Compare Dask shape to computed result shape
As an extra check of the shape, make sure the shape hasn't changed
between the Dask and the computed result. Have seen this occur in some
cases as well while working on other code. So this should show some of
these issues and help avoid them cropping up in the future.