Fix handling of *rfft style functions
Pass around the expected result type of the functions to all of the
chunking operations. Use this information to determine the proper chunk
sizes particularly for functions like `rfft` and `irfft`, which differ
between NumPy and SciPy. However this comes with one caveat, `irfft`
will be dispatched according to whether the data is complex (NumPy) or
real (SciPy).