Abstract FFT interface (#2093)
* Restructure FFT wrapper for custom wrapping

Change `_fft_wrap` to delay function specification. The goal being one
can create custom wrappers for different external FFT implementations.
This both keeps Dask flexible in the area of FFTs without necessarily
adding many implementations.

* Expose FFT wrappers in the public API and use them

Create public API FFT wrapper functions and make use of them to wrap the
exist NumPy FFTPACK interface used by Dask. Also provide some brief
explanation of how this API works in the docstrings.

* Test explicit wrapping NumPy's FFTPACK

Even though we are already doing this wrapping internally in Dask and
implicitly testing. We should make sure explicitly that it works in case
the implementation in Dask changes in the future.

* Test wrapping SciPy's FFTPACK

As another exercise of the wrappers, test wrapping SciPy's FFTPACK to
make sure our wrappers apply just as well to it.

* Consolidate to one fft_wrapper function

Instead of having a variety `*fft_wrapper` functions, simply provide one
`fft_wrapper` function that takes a specification as to the kind of FFT
function that is being wrapped.

* Parameterize wrapper tests to consolidate them

To cutdown on the amount of boilerplate in the tests, use
`pytest.mark.parametrize` so as to cycle through many different
combinations. This should keep the code small while allowing us to try
various combinations of NumPy and SciPy FFT functions with the wrappers.

* Raise unknown kind provided to `fft_wrapper`

If an unknown kind is provided to `fft_wrapper`, raise a `ValueError`
explaining that this happened.

* Handle FFT wrapping kinds with a dictionary

To simplify passing parameters to the wrapper, create a dictionary that
has all of the parameters for each supported kind so that they can be
easily passed through.

* Document the wrapper_fft API better

Explain the function's expected API more clearly. Also note the
supported kinds of functions that can be wrapped and how they should be
specified.

* Consolidate remaining FFT wrapping tests

Combine the remaining NumPy tests into the single parameterized test for
FFT wrappers. Also as SciPy's FFTPACK implements `rfft` and `irfft` in
ways that are not compatible with NumPy's API, skip testing the wrapping
of the functions and note why they are skipped. Should keep testing
relatively straightforward.

* clean up fft_wrapper

* support more scipy.fft functions

* clean up map_blocks call

This removes the call to partial

* Remove unnecessary indiretion with _fft_wrap

* stop using kind

* update style

* make out_chunk_fns private

* Test a different `n` with the wrappers

Just to make sure we don't fall victim to some faulty assumptions, test
a different `n` to see if the wrappers still do the right thing compared
to the non-wrapped versions.

* Fill out wrapped FFTs unknown type based on array

Determines the unknown `dtype` based on using a dummy array of the same
`dtype` as that of the array provided. That way, we are sure it should
match the result's `dtype` accurately based on the type we are provided.

* Hack around nonlocal variable issue

Seems the closure doesn't bind the argument from the outer scope
function. So hack around this issue by storing it to a variable that is
local to the outer function and then copy it over to the inner function.
Despite being quite ugly, this does seem to work correctly.

* Add some FFT wrapper tests for single precision

* Parameterize FFT wrapper type testing

To make sure types are properly handled throughout, try converting the
data to single and double precision before proceeding.

* Use assert to compare types from FFT wrappers

Was using `assert_eq`, which is geared for arrays. So fixed this to do a
simple equal comparison in an `assert`.

* Test that FFT wrapper shape matches actual shape

Just to make sure the correct shape is computed for the Dask array, add
a test that verifies that the shape of the Dask after the FFT matches
that of the actual result.

* Fix handling of *rfft style functions

Pass around the expected result type of the functions to all of the
chunking operations. Use this information to determine the proper chunk
sizes particularly for functions like `rfft` and `irfft`, which differ
between NumPy and SciPy. However this comes with one caveat, `irfft`
will be dispatched according to whether the data is complex (NumPy) or
real (SciPy).

* Fix some linting issues

* Revert "Fix handling of *rfft style functions"

This reverts commit 4257a2bb64b5d5e6215e3d3d0b6abcc690901302.

* Error on SciPy's `rfft` and `irfft` functions