Add FFTs for 2 and N dimensions
This expands on `fft_wrap` to allow support for wrapping functions of
the form `*fft2` and `*fftn`. To provide this support separate chunk
generating functions were added to handle both the 2-D and N-D cases
together.

In addition to providing this support, the NumPy functions for 2-D and
N-D cases that match the supported API were also wrapped. Though it
should be possible to wrap other FFT functions from external libraries
(as long as the support the NumPy API) should that be desired.