Add Multidimensional FFTs (#2116)
* Add FFTs for 2 and N dimensions

This expands on `fft_wrap` to allow support for wrapping functions of
the form `*fft2` and `*fftn`. To provide this support separate chunk
generating functions were added to handle both the 2-D and N-D cases
together.

In addition to providing this support, the NumPy functions for 2-D and
N-D cases that match the supported API were also wrapped. Though it
should be possible to wrap other FFT functions from external libraries
(as long as the support the NumPy API) should that be desired.

* Consolidate FFT chunk generating functions

To simplify 1-D and N-D FFT support, this combines the chunk generating
functions for both cases and makes some tweaks to the 1-D case to treat
it as a special case of N-D. Should make the code a little more compact
and hopefully a bit easier to test.

* Consolidate 1-D and N-D functions

Instead of having separate wrappers for 1-D and N-D functions, add one
wrapper function for both cases. To resolve occasional discrepancies
between the functions add a few branches to handle the different
dimension cases. Also add a small additional wrapper for the 1-D case to
map and format arguments from the 1-D case into something the N-D case
can handle.

* Test multidimensional FFTs

Adds various tests to exercise 2-D and N-D FFT functions. Also tests
wrapping of functions from NumPy and SciPy with 2-D and N-D variants.

* Disallow repeated axes to FFT functions

As it is sort of strange for someone to take an FFT of an FFT, simply
disallow this behavior by raising an exception. Also check that this
exception is raised.

* Tweak `test_fft2n` to include another shape case

* Add better names for the N-D FFT tests

* Remove excess dim, dtype variables

This is just one less thing to track.  Instead we inline the kind check.

* Assign _dtype and fix indentation