Improve `fuse` performance by allowing tasks to reference old key names that are now aliases. (#2129)
Previously, tasks that depend on fused tasks whose keys were renamed would be
updated via `subs` to use the new task keys.  In some cases, the number of tasks
that depend on fused tasks would greatly outnumber the number of fused tasks,
and calling `subs` on them would dominate the calculation time.

Now, we don't call `subs` on tasks that depend on renamed fused tasks.  These
tasks are no longer modified and use the previous key name, which is now an
alias, `{old_key: new_key}`.  This can result in a larger graph and slightly
more scheduler overhead, but the performance increase to `fuse` can be significant.