Add a roll function for Dask Arrays
Provides a function that performs the same operation as NumPy's roll
except on Dask Arrays. This is done by using a mixture of slicing and
concatenating. The effect is a Dask Array that is either raveled,
rolled, and reshaped or rolled along specified axes. Added a few tests
and added the function to the API doc.