Check and compare shapes in assert_eq (#2101)
* Check and compare shapes in assert_eq

Seems that shape comparison only occurred on the result of computing the
Dask array. In our case, we found the Dask array itself had the wrong
shape, but the computed result had the right shape. To fix this we add a
comparison of the shapes before computation occurs. This happens in
addition to the existing test comparing the shapes afterwards.

* Compare Dask shape to computed result shape

As an extra check of the shape, make sure the shape hasn't changed
between the Dask and the computed result. Have seen this occur in some
cases as well while working on other code. So this should show some of
these issues and help avoid them cropping up in the future.

* check dask.array shape against computed shape