`dask.threaded.get` can be interrupted in python 2
Due to a bug in `Queue.get` in python 2, `get` is not interruptible if
no timeout is set (see https://bugs.python.org/issue1360). To get around
this, we specify an extremely long timeout. Benchmarking shows no
noticeable overhead to specifying a timeout, and this allows
`dask.threaded.get` to be interrupted.