Test rolling 2 axes at once (#2141)
* Actually raise the TypeError, duh!

* Ensure roll has the same number of shifts as axes

There shouldn't be more shifts than axes or vice versa. So this check
ensures there are the same number of each and raises a `ValueError` if
the constraint is not met.

* Test rolling 2 axes at once

Make sure that `roll` still works correctly on Dask Arrays when rolling
2 dimensions at once. This is done in addition to testing 1 dimension
rolling.

* Consolidate 1-D and 2-D roll tests

Instead of having a separate set of 2-D roll tests, combine them with
the 1-D roll tests. Make sure proper exceptions are raised for
mismatches and handled accordingly as well.