`dask.threaded.get` can be interrupted in python 2 (#2144)
* `dask.threaded.get` can be interrupted in python 2

Due to a bug in `Queue.get` in python 2, `get` is not interruptible if
no timeout is set (see https://bugs.python.org/issue1360). To get around
this, we specify an extremely long timeout. Benchmarking shows no
noticeable overhead to specifying a timeout, and this allows
`dask.threaded.get` to be interrupted.

* More robust interrupt handling

- Properly handle windows with python 3
- Add test for keyboard interrupt

* Fix test for windows

* Remove slow mark on interrupt test

* Re-skip the test on appveyor