Add sql loader (#1181)
* Add sql loader

* Start attempt to use ntile instead of offsets

This is untestable as ntile and row_number are not implemented
in sqlite (but in the target sql servers, e.g., PG, they are).

* Tests pass on postgres

* Explicitly set index

* Remove dross

* Try simple SQL->dataframe loader

* Flake fixes

* Now with SQLalchemy

The code looks a little less tidy, and I don't like the isinstance(, str), but
seems to work well for a range of cases.

* Clean up

* flake fixes

* Requested fixes

Also, added some tests and uncovered some corner cases.

Not insisting that, if the index is an sqlalchemy compountd construct,
it should be explicitly named using `.label()`. Raises exception
otherwise. This gets around trying to guess what pandas will name
that column.

* Further fixes

Split out npartitions and divisions to avoid confusion.

* cosmetics

* Trailing u

* Furhter small changes

* Make upper limit inclusive

Similifies what happens with max value

* Create SQL engine within task

Tests only worked on multi-threaded, because the sqlalchemy engine is not
serializable.

* flake

* Auto-guess chunksize if npartition not given

Also: enable operating on any input sqlalchemy expression, not only
a named table.

* more flake

* add read_sql_table to imports and docs

* Ensure time divisions are are range limits