No longer ignore the FFT tests
It appears the FFT tests were being explicitly skipped by this
configuration file. A look into the commit history suggests this had
something to do with avoiding doctests. Though it is unclear whether
there were any doctests. In any event, there are none now and this is
unintentionally suppressing the unittests. So simply remove this
configuration file as it didn't do anything else.