Pin NumPy and pandas in the environment
Use the `pinned` file for the test environment to fix the versions of
NumPy and pandas. This should ensure these don't need to be explicitly
listed in each step as these pins will always be matched no matter what.
Also by doing this, there is less opportunity of forgetting to
explicitly fix these and thus run into issues.