Specify the NumPy patch version on AppVeyor
Appears that NumPy 1.12.1 has been used on AppVeyor recently. So picking
that patch version seems reasonable. Given we pin to the patch version
with pandas on AppVeyor, this doesn't seem unreasonable to do with
NumPy. Further on Travis CI we also pin to the patch version with NumPy.
In short, this is a totally reasonable and consistent change, which
should have minimal effect.