Run the FFT tests and fix a few failing tests (#2183)
* No longer ignore the FFT tests

It appears the FFT tests were being explicitly skipped by this
configuration file. A look into the commit history suggests this had
something to do with avoiding doctests. Though it is unclear whether
there were any doctests. In any event, there are none now and this is
unintentionally suppressing the unittests. So simply remove this
configuration file as it didn't do anything else.

* Enumerate the axes to properly order the chunks

The output shape was being incorrectly ordered in some cases as the
shape was remapped by the axis order, but the axes themselves. This
corrects that by enumerating the axes and using the enumeration to
reorder the output shape for use with chunks.

* Adjust handling of 1-D FFTs

Given the changes to how we handle the output shape in the FFTs, this
breaks the 1-D FFT special case code. So update it to reflect how the
output shape is now being handled.

* Fix output shape handling in real FFTs

* Re-add conftest so as to exclude fft doctests only

Seems the wrapped functions in the fft module expose docstrings with
doctests, which pytest picks up on and tries to run. So re-add the
existing configuration file, but update it to exclude the file `fft.py`
in `dask/array` only.