Fix CI pinnings (#2185)
* Pin NumPy and pandas in the environment

Use the `pinned` file for the test environment to fix the versions of
NumPy and pandas. This should ensure these don't need to be explicitly
listed in each step as these pins will always be matched no matter what.
Also by doing this, there is less opportunity of forgetting to
explicitly fix these and thus run into issues.

* Drop pinning of NumPy and pandas during install

These pinnings are now set in the pinning file. So we merely need to
install NumPy and pandas once and they will stick to the same version.

* Specify the NumPy patch version on AppVeyor

Appears that NumPy 1.12.1 has been used on AppVeyor recently. So picking
that patch version seems reasonable. Given we pin to the patch version
with pandas on AppVeyor, this doesn't seem unreasonable to do with
NumPy. Further on Travis CI we also pin to the patch version with NumPy.
In short, this is a totally reasonable and consistent change, which
should have minimal effect.