Minor Cleanup of 1D FFTs (#2216)
* For 1-D FFTs, handle unknown `axes` in the wrapper

Was filing in unknown `axes` for the 1-D case outside of the wrapper
because of some old code that relied on knowing that information
earlier. As we no longer need this information earlier, just handle
unknown `axes` in the wrapper as we already do for 2-D and N-D cases.
Should also drop an untested line in the process.

* With 1-D FFTs, make `s` a `tuple` if `n` is known

Was making `s` a `list` when `n` was known before, this makes it a
`tuple`, which is more consistent with `s` in all of the other cases.