Fix map_overlap for trivial padding
In `add_dummy_padding` there was a special workaround for the `"none"`
boundary condition that failed when the `depth` was trivial (i.e. `0`).
To fix this, simply check to see that the boundary condition is `"none"`
and the `depth` is positive-definite before proceeding. This appears to
clear up the issue when working trivial depth and the `"none"` boundary
condition.