Skip rechunking after calling repeat
It appears that calling `repeat` here results in a chunk size that is
equal to the number of repetitions. As such these repeated dimensions
will be rechunked to the right chunk size automatically. Namely
rechunking will be done according to the chunk size of all of the
`arange` calls. However when it comes to `stack` rechunking is still
required to return the intended chunk size.