Skip chunking of 0th dimension in indices
The `indices` function was chunking the 0th dimension based on the user
supplied `chunks` argument. Following some discussion it was decided
that chunking along this dimension should simply be skipped. After all
the user can easily do this themselves if they want. Plus this
simplifies a bit of code in the process. So now the chunks are only used
for the `arange` function calls. Thus it determines the 1st dimension
onwards. The 0th dimension keeps chunks of length 1.