Add indices (#2268)
* Add an indices function

Behaves like NumPy's equivalently named function except with Dask Arrays
instead.

* Include some tests for the new indices function

* Skip rechunking after calling repeat

It appears that calling `repeat` here results in a chunk size that is
equal to the number of repetitions. As such these repeated dimensions
will be rechunked to the right chunk size automatically. Namely
rechunking will be done according to the chunk size of all of the
`arange` calls. However when it comes to `stack` rechunking is still
required to return the intended chunk size.

* Simply import itertools

Use `chain` from the module instead of importing the function directly.

* Skip chunking of 0th dimension in indices

The `indices` function was chunking the 0th dimension based on the user
supplied `chunks` argument. Following some discussion it was decided
that chunking along this dimension should simply be skipped. After all
the user can easily do this themselves if they want. Plus this
simplifies a bit of code in the process. So now the chunks are only used
for the `arange` function calls. Thus it determines the 1st dimension
onwards. The 0th dimension keeps chunks of length 1.

* Document the indices function