Updated runtime requirements and changes in using the lz4 compression api (#1032)
* Update compression.py

updated lz4 support to no longer use the deprecated lz4.LZ4_*compress functions.

* Update requirements.txt

Added bokeh (for distributed status page) and paramiko (for dask-ssh) to requirements

* Update requirements.txt

Removed bokeh and paramiko from the runtime requirements as they are considered to be optional. bokeh is not strictly needed to use distributed and paramiko is only necessary if dask-ssh is used to start a distributed cluster.

* Update compression.py

Added support to distinguish between older and new versions of lz4, as we cannot expect to have the latest lz4 packages available in older python environments.

* Update compression.py

fixed indentation error.

* Update compression.py

fixed some style issues.