Check for tricky cases when fusing fancy indices
Previously we'd fuse normal and fancy indexing like:

```python
a[:, :, [0, 1], :][0, :, :, :] -> a[0, :, [0, 1], :]
```

This is incorrect, and was leading to bugs:

```python
>>> a = np.ones((11, 13, 20, 12))
>>> a[:, :, [0, 1], :][0, :, :, :].shape
(13, 2, 12)
>>> a[0, :, [0, 1], :].shape
(2, 13, 12)
```

Let `ind1` and `ind2` represent two index operations as in
`x[ind1][ind2]`. We now only fuse indices involving fancy indexing if:
- Only one of the two uses fancy indexing
- Dimensions where fancy indexing isn't used are non-integers

This means that the following will be fused:

```
a[:10, [1, 2, 3]][:5, 0]
```

but this won't:

```
a[:10, [1, 2, 3][0, 0]
```

This does exclude some valid cases, but is easier to write than
following all the checks required when mixing fancy and non-fancy
indexing.