Indexing with empty list gives correct chunks
Previously indexing with an empty list would result in chunks with empty
tuples, leading to x.numblocks reporting 0 along the indexed dimension.
Since `x[[]]` and `x[:0]` are equivalent when there's only one list
(which is all we support), we fallback to normal slicing behavior when
an empty list is encountered.