Add rfftfreq (#2322)
* Add an rfftfreq function for Dask Arrays

Much like the `fftfreq` function added for Dask Arrays, the `rfftfreq`
function leverages `linspace` to get an appropriate range of values.
However as the whole range is not needed in this case and nothing else
needs to be done to elements of the sequence beforehand, `rfftfreq`
merely truncates the sequence at the intended endpoint and proceeds to
handle the inverse scaling factor before returning the result.

* Refactor out a helper function for [r]fftfreq

As there is a lot of similarity between the `rfftfreq` and `fftfreq`
functions, this refactors out a common more general function that
applies to both use cases. Should help reduce the amount of code
involved and keep things maintainable.

* Reuse fftfreq tests for rfftfreq as well

As we are going to want to test the same sorts of things for both
`fftfreq` and `rfftfreq`, simply parameterize the `fftfreq` tests to
take a function name so that `fftfreq` and `rfftfreq` can be run through
the same tests.