Handle extra chunk in *fftfreq functions
Chunking is not handled correctly with the various *fftfreq. There seems
to be some subtle edge cases due to the fact that we construct arrays
with the chunks that are larger than the final result. To fix this
generally we have to normalize the chunks and tack on additional chunks
for the portion that will be disposed.

In the case of `fftfreq` and `ifftfreq`, only one extra value is added
by `linspace`. To adjust the provided chunks for this, we tack on a
singleton chunk to the end after normalization. This should allow it to
be trivial stripped off once we are done as it has no overlap with
anything else. So this shouldn't have any performance overhead.

As for the case of `rfftfreq`, we have to do a bit more. In this case,
we strip off about half of the `linspace` we construct and drop it. To
fix this case we tack on an additional chunk that is the length of this
pieces that will be dropped. As the chunks will be normalized again,
this last chunk will be probably be split into similar sizes as the
other chunks. Also the last chunk will probably have some overlap with
a portion of `linspace` that is to be removed. This might have some
effect on performance, but shouldn't be negligible when there are many
chunks and/or chunks are small.