Always create worker's local directory (#1079)
* Always create worker's local directory

Previously if given a directory we put files directly into that
directory.  This caused problems because users gave us large spaces like
/tmp or /local.  It also required us to track if we wanted to delete the
directory afterwards or not.

Now we *always* create a new directory and we always delete it when
finished.  This removes the ability to specify a particular space, but
it's not clear that this was being used.  This seems to be more
predictable and less error prone.

Also, we remove the temp_filename keyword, which is no longer needed.

* flake8