Spring cleaning (#2346)
* Delete unused definitions in dask.array

* Remove unused definitions in dask.bag

* Remove unused definitions in dask.bytes

* Remove unused definitions in dask.delayed

* Remove deprecated dask.delayed functions

Removed:
- dask.delayed.do
- dask.delayed.compute
- dask.do

* Remove unused definitions in dask.threaded

* Remove unused definitions in dask.dataframe.utils

* Remove unused definitions in dask.utils

* Remove unnecessary `eq_strict` function

Move `no_result` to be a singleton object, use `is` instead of
`eq_strict`

* Move some utils -> bytes.utils

Move functionality in `dask.utils` that was only/mainly used by bytes to
`bytes.utils`

* Move infer_storage_options definition to dask.utils

Distributed imports `dask.utils.infer_storage_options`. Since we want
dask to work with older versions of distributed, we need to keep that
import around for a while. We can't define it in `dask.bytes.utils`
since that results in circular imports. As such, we keep the definition
in `dask.utils`, import it in `dask.bytes.utils`, and always use the
later internally. Once distributed's dask requirements have been bumped
we can move the definition.