Add docpage for external packages
We may want a page like this to increase exposure to other packages that use or
build on Dask.  The goal here is that when someone says "How do I use Dask to
solve X kind of problem" that we can send them to a subsection of this page for
a set of links to other downstream projects.

Creating such a comprehensive list is hard.  This could use help from others.