Some async-await friendly changes (#1115)
* Support async def coroutines in gen_cluster

* Remove start= keyword from client

Previously when operating asynchronously we would have to explicitly
pass in start=False and then `yield client._start()`.  This was
confusing for novice users and looked unpleasant.

Now we detect if we are inside of an IOLoop or not by checking the
IOLoop's thread and our thread.  Additionally we make the client itself
awaitable so that the following syntax becomes the basic way of starting
a client asynchronously

```
client = Client(..., start=False)
await client._start()
```

```
client = await Client(...)
```

However we are now relying on internal Tornado API to get the thread
identity of the IOLoop

* remove tornado 4.5 dependence

* rename test file

Otherwise pytest triggers even on python 2

* create separate in_ioloop function

* avoid reusing current loop after fork

* Replace `yield future._result()` with `yield future`