Forget errant entries in who_has after comm failure (#1121)
Workers keep entries about which of their peers have which data.  When a worker
goes down we should update this structure to remove stale entries.

In order to test this we needed the ability to consistently remove dependencies
from a worker, which led to workers informing the scheduler when the clear out
old data.