Remove nans from divisions when shuffling
When shuffling we first check to see if the column is already sorted, in which
case we can be much much faster.  To do this we compute the minimum and maximum
value of the column in every partition.  If there is an empty partition then
this can cause a problem, because the min/max value will be nan.

Now we remove these nans before proceeding

Fixes https://github.com/dask/dask/issues/2408