Optimize out broadcast_to on unchanged shape (#2417)
* Optimize out `broadcast_to` on identical shape

If `broadcast_to` is called with a shape that is already equal to the
Dask Array's current shape, simply return the array unchanged. There is
no need to do extra work if the condition is already met.

* Test broadcast_to optimization for same shape

Verify that Dask Array's identity is unaltered between input and return
value if the shape is unchanged.