Fixes some type issues in Dask Array shapes (#2375)
* Handle depth of other integral types

Instead of requiring that `depth` be of type `int`, allow `depth` to be
of any integral type. Also handle all integral types provided for
`depth` the same way.

* Reconstruct `depth` dictionary

If a dictionary for `depth` is provided (or formed), handle conversion
of the keys and values to `int` type. This is done so as to avoid weird
integral types showing up in the shape of Dask Arrays.

* Drop blank line

* Test shape value types after `map_overlap`

Make sure that even if the `depth` provided to `map_overlap` has a
different type from `int`, that `map_overlap` proceeds through and
properly converts the Integral values to `int` instances.

* normalize chunk values to integer type