Cleanup and fortify where tests (#2428)
* Test `where` on mismatched types

In order to make sure type promotion occurs correctly, have chosen the
two arrays to have different types. Should be promoted to the largest of
the two types.

* Test `where` with pure Boolean conditions

Verify that `where` is able to behave correctly with pure Boolean
conditions even with optimizations applied.

* Refactor out another test loop

* Break up assertion check in where

* Use `astype` with `randint`

It appears that in NumPy 1.10, `randint` does not support a `dtype`
argument. Fortunately it is a simple matter to convert with `astype`
after the fact.

* Test Python and NumPy bools in where

Make sure to test both Python and NumPy bools in `test_where` to make
sure they are both handled correctly. This is important check when
optimizations are added for these cases.