Optimize scalar condition in where (#2432)
* Optimize Dask Array's where with scalar condition

If the `condition` provided to Dask Array's `where` function is a simple
scalar value (i.e. a Boolean), simply pick one of the two arrays and
return it. Also make sure to promote the returned Array's type and
broadcast the Array's shape to be consistent both with Dask's previous
behavior and NumPy's behavior.

* Test where for optimization

Add some tests to verify the resulting array from `where` has the same
identity as the input array if a `bool` value is provided as the
condition.