Slice delayed (#2467)
* delayed works with slicing

Previously we weren't pulling delayed objects out of `slice` objects,
causing slicing to fail on delayed objects. This fixes that.

* Fix xfailed delayed Iterator test

This was testing for behavior that has never been implemented, and
probably can't be. We fix the test to check that passing iterators to
delayed functions works as currently implemented (converts iterator to
a list).